statistics: split lock stats handler and rewrite the insert SQL
statistics: update SQL

statistics: add tests
