+++ statistics/handle/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
         "handle.go",
         "handle_hist.go",
         "historical_stats_handler.go",
+        "lock_stats_handler.go",
         "update.go",
     ],
     importpath = "github.com/pingcap/tidb/statistics/handle",
@@ -59,13 +60,14 @@ go_test(
         "dump_test.go",
         "gc_test.go",
         "handle_hist_test.go",
+        "lock_stats_handler_test.go",
         "main_test.go",
         "update_list_test.go",
     ],
     embed = [":handle"],
     flaky = True,
     race = "on",
-    shard_count = 26,
+    shard_count = 27,
     deps = [
         "//config",
         "//domain",
